チームとプロジェクトのアイテムコンポーネントを共通化
