+++ iruka-code/src/app/(main)/_components/project-item.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import { cn } from '@/lib/utils';
 import { Id } from '@convex/_generated/dataModel';
-import { ChevronDown, ChevronRight, LucideIcon, Plus } from 'lucide-react';
+import { LucideIcon } from 'lucide-react';
 import { useRouter } from 'next/navigation';
+import { Item } from './item';
 
 interface ItemProps {
   id?: Id<'projects'>;
@@ -30,52 +30,19 @@ export const ProjectItem = ({
 }: ItemProps) => {
   const router = useRouter();
 
-  const handleExpand = (
-    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
-  ) => {
-    event.stopPropagation();
-    onExpand?.();
-  };
-
-  const ChevronIcon = expanded ? ChevronDown : ChevronRight;
-
   return (
-    <div
-      onClick={onClick}
-      role="button"
-      style={{ paddingLeft: level ? `${level * 12 + 12}px` : '12px' }}
-      className={cn(
-        'group min-h-[27px] text-sm py-1 pr-3 w-full hover:bg-primary/5 flex items-center text-muted-foreground font-medium',
-        active && 'bg-primary/5 text-primary',
-      )}
-    >
-      {!!id && (
-        <div
-          role="button"
-          className="h-full rounded-sm hover:bg-neutral-300 dark:bg-neutral-600 mr-1"
-          onClick={handleExpand}
-        >
-          <ChevronIcon className="h-4 w-4 shrink-0 text-muted-foreground/50" />
-        </div>
-      )}
-      <Icon className="shrink-0 h-[18px] mr-2 text-muted-foreground" />
-      <span className="truncate">{label}</span>
-      {isSearch && (
-        <kbd className="ml-auto pointer-events-none inline-flex h-5 select-none items-center gap-1 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium text-muted-foreground opacity-100">
-          <span className="text-x">⌘</span>K
-        </kbd>
-      )}
-      {!!id && (
-        <div className="ml-auto flex items-center gap-x-2">
-          <div
-            role="button"
-            onClick={() => {}}
-            className="opacity-0 group-hover:opacity-100 h-full ml-auto rounded-sm hover:bg-neutral-300 dark:hover:bg-neutral-600"
-          >
-            <Plus className="h-4 w-4 text-muted-foreground" />
-          </div>
-        </div>
-      )}
-    </div>
+    <Item
+      {...{
+        id,
+        label,
+        onClick,
+        icon: Icon,
+        active,
+        isSearch,
+        level,
+        onExpand,
+        expanded,
+      }}
+    />
   );
 };
